+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.0.7] - 2023-04-26
+### Added
+- Add method to EnumHelper to get a random value in a specific range
+
+## [1.0.6] - 2023-03-10
+### Fix
+- Fix EnumDictionary foldout & position.
+
+
+## [1.0.5] - 2023-02-23
+### Fix
+- Remove Task delegate because useless and was preventing System.Threading.Tasks.Task to be used in FishingCactus without specifying the whole path.
+
 ## [1.0.4] - 2022-01-27
 ### Added
 - Add Max attribute for integer and float fields and Optional helper struct